+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Group changes to describe their impact on the project, as follows:
 
 ## [1.1.0] - Alpha
 
+### [Fixed]
+- Call logs are disabled to avoid memory accumulation
+
 ## [1.0.0] - Beta
 The following changes are made in comparison to the conference server released in Flexisip 2.5.1.